+++ src/pytest_ranking/plugin_utils.py
@@ -5,3 +5,4 @@ DATA_DIR = "pytest_ranking_data"
 # Default amount of historical test run results to store per test
 DEFAULT_HIST_LEN = 50
 DEFAULT_WEIGHT = "1-0-0"
+DEFAULT_SEED = 0