+++ CHANGELOG.md
@@ -6,9 +6,10 @@
 
 * Add textual similarity between tests and changed files since last run as the third heuristic
 * Weight normalization
+* History length as an optional argument
 
 
 0.1.0 (2023-12-04)
 ----
 
-* First release on PyPI.
+* First release on PyPI